third commit

+++ src/models/orderModel.ts
@@ -9,14 +9,21 @@ type Order = {
   status: string;
 };
 
-exports.getAllOrders = async (
-  _req: Request,
-  res: Response
-): Promise<Order[]> => {
+exports.getAllOrders = async (req: Request, res: Response) => {
+  const status = req.query.status;
+  const currentUser = res.locals.user;
+  const user_id = currentUser.user_id;
+  console.log(user_id, status);
   try {
+    let result;
     const conn = await client.connect();
-    const sql = 'SELECT * FROM orders';
-    const result = await conn.query(sql);
+    if (status && status === ('active' || 'completed')) {
+      const sql = `SELECT * FROM orders WHERE user_id=($1) AND status=($2)`;
+      result = await conn.query(sql, [user_id, status]);
+    } else {
+      const sql = `SELECT * FROM orders WHERE user_id=($1)`;
+      result = await conn.query(sql, [user_id]);
+    }
     conn.release();
     res.status(200).json({
       status: 'success',
@@ -28,16 +35,14 @@ exports.getAllOrders = async (
     return result.rows;
   } catch (err) {
     res.status(400).json(err);
-    throw new Error(`Cannot get all orders ${err}`);
+    console.error(err);
   }
 };
 
-exports.createOrder = async (req: Request, res: Response): Promise<Order> => {
+exports.createOrder = async (req: Request, res: Response) => {
   const currentUser = res.locals.user;
   const user_id = currentUser.user_id;
-  console.log(currentUser);
   const { product_id, quantity, status } = req.body;
-  // console.log(product_id, quantity, user_id, status);
   try {
     const conn = await client.connect();
     const sql = `INSERT INTO orders (product_id, quantity, user_id, status) VALUES ($1, $2, $3, $4) RETURNING *`;
@@ -57,37 +62,41 @@ exports.createOrder = async (req: Request, res: Response): Promise<Order> => {
     return result.rows[0];
   } catch (err) {
     res.status(400).json(err);
-    throw new Error(`Cannot create an order ${err}`);
+    console.error(err);
   }
 };
 
-exports.getOrder = async (req: Request, res: Response): Promise<Order[]> => {
-  const order_id = req.params.id;
-  try {
-    const conn = await client.connect();
-    const sql = `SELECT * FROM orders WHERE order_id=($1)`;
-    const result = await conn.query(sql, [order_id]);
-    conn.release();
-    // console.log(order);
-    res.status(200).json({
-      status: 'success',
-      data: {
-        order: result.rows[0],
-      },
-    });
-    return result.rows[0];
-  } catch (err) {
-    res.status(400).json(err);
-    throw new Error(`Cannot get an order ${err}`);
-  }
-};
+// exports.getOrder = async (req: Request, res: Response) => {
+//   const currentUser = res.locals.user;
+//   const user_id = currentUser.user_id;
+//   const order_id = req.params.id;
+//   try {
+//     const conn = await client.connect();
+//     const sql = `SELECT * FROM orders WHERE order_id=($1) AND user_id=($2)`;
+//     const result = await conn.query(sql, [order_id, user_id]);
+//     conn.release();
+//     // console.log(order);
+//     res.status(200).json({
+//       status: 'success',
+//       data: {
+//         order: result.rows[0],
+//       },
+//     });
+//     return result.rows[0];
+//   } catch (err) {
+//     res.status(400).json(err);
+//     throw new Error(`Cannot get an order ${err}`);
+//   }
+// };
 
-exports.deleteOrder = async (req: Request, res: Response): Promise<Order[]> => {
+exports.deleteOrder = async (req: Request, res: Response) => {
+  const currentUser = res.locals.user;
+  const user_id = currentUser.user_id;
   const order_id = req.params.id;
   try {
     const conn = await client.connect();
-    const sql = `DELETE FROM orders WHERE order_id=($1) RETURNING *`;
-    const result = await conn.query(sql, [order_id]);
+    const sql = `DELETE FROM orders WHERE order_id=($1) AND user_id=($2) RETURNING *`;
+    const result = await conn.query(sql, [order_id, user_id]);
     const deletedOrder = result.rows[0];
     conn.release();
     res.status(204).json({
@@ -96,24 +105,26 @@ exports.deleteOrder = async (req: Request, res: Response): Promise<Order[]> => {
     });
     return deletedOrder;
   } catch (err) {
-    throw new Error(`Cannot delete the order(order_id: ${order_id}) ${err}`);
+    res.status(400).json(err);
+    console.error(err);
   }
 };
 
-exports.updateOrder = async (req: Request, res: Response): Promise<Order[]> => {
-  const { product_id, quantity, user_id, status } = req.body;
+exports.updateOrder = async (req: Request, res: Response) => {
+  const { product_id, quantity, status } = req.body;
   const order_id = req.params.id;
-  req.body;
+  const currentUser = res.locals.user;
+  const user_id = currentUser.user_id;
   try {
     const conn = await client.connect();
     const sql =
-      'UPDATE orders SET product_id=($1), quantity=($2), user_id=($3), status=($4) WHERE order_id=($5) RETURNING *';
+      'UPDATE orders SET product_id=($1), quantity=($2), status=($3) WHERE order_id=($4) AND user_id=($5), RETURNING *';
     const result = await conn.query(sql, [
       product_id,
       quantity,
-      user_id,
       status,
       order_id,
+      user_id,
     ]);
     conn.release();
     res.status(200).json({
@@ -124,6 +135,7 @@ exports.updateOrder = async (req: Request, res: Response): Promise<Order[]> => {
     });
     return result.rows[0];
   } catch (err) {
-    throw new Error(`Cannot update the order(order_id: ${order_id}) ${err}`);
+    res.status(400).json(err);
+    console.error(err);
   }
 };
